+++ translations/src/main/kotlin/com/hedvig/libs/translations/TranslationsClient.kt
@@ -1,6 +1,8 @@
 package com.hedvig.libs.translations
 
+import java.util.Locale
+
 interface TranslationsClient {
-    fun getTranslation(locale: String, key: String): String?
+    fun getTranslation(locale: Locale, key: String): String?
 }
